fix(core): minor tweak to config error string (#414)
Small improvement to the config error when didn't provide a definition
for every variable in a contract. Was unclear if the message was
referring to a variable that was in the contract but not in the config
or if the variable was in the config but not in the contract.

Co-authored-by: sam-goldman <106038229+sam-goldman@users.noreply.github.com>